Add restart from snapshot test and fix snapshot handling issues
- Add comprehensive test case for restarting nodes from snapshots
- Fix snapshot handling in Raft implementation
- Move snapshot index setting to proper location
- Simplify gRPC network connections
- Remove redundant mock loggers from tests

+++ test/testcluster/babuza_integration_test.go
@@ -725,7 +725,7 @@ func TestBabuza_Cluster_ClientRegisterSession(t *testing.T) {
 				bootstrap, err := defaultBootstrapBuilder(&appConfig.BubuzaConfig, babuzaDirs,
 					appConfig.VotingPeersCfg, pn, babuzaLogger)
 				assert.Nil(t, err)
-				bootstrap.SetSessionManager(session.NewExpiredManager(babuzaLogger, session.SetExpiredMgrOptionsWithExpiredTime(int64(sessionExpiredDuration))))
+				bootstrap.SetSessionManager(session.NewExpiredManager(babuzaLogger, session.SetExpiredMgrOptionsWithExpiredTime(sessionExpiredDuration)))
 				return CreateKvEmbeddedApp(appConfig, kvstore.NewMemoryStoreWithSession(), bootstrap)
 			})
 		wait := tc.RaftElectionTimeout() * 3
